Remove sidekiq-unique-jobs
There are too many issues with keys staying stuck forever.
At this moment, while long polling is not supported, we don't
need it anyway. When we add long polling support, we will have
to have a more active locking mechanism, wherein the worker, as
its doing its work, can report that it's alive and still working
before the lock expires